ISS73 - refactor simulation exec
