+++ python_src/models/__init__.py
@@ -7,4 +7,5 @@ from .ac3r import CrashScenario
 from .simulation_factory import SimulationFactory
 from .simulation import Simulation
 from .simulation_score import SimulationScore
+from .simulation_execution import SimulationExec
 from .kmeans import KMeans
